perf: use high-resolution danmaku timing

+++ lib/base_danmaku_painter.dart
@@ -8,10 +8,10 @@ abstract base class BaseDanmakuPainter extends CustomPainter {
   final double strokeWidth;
   final bool running;
   final int batchThreshold;
-  final int _tick;
-  final ValueListenable<int>? repaint;
+  final double _tick;
+  final ValueListenable<double>? repaint;
 
-  int get tick => repaint?.value ?? _tick;
+  double get tick => repaint?.value ?? _tick;
 
   static final Paint _paint = Paint();
 
@@ -21,7 +21,7 @@ abstract base class BaseDanmakuPainter extends CustomPainter {
     required this.fontWeight,
     required this.strokeWidth,
     required this.running,
-    required int tick,
+    required double tick,
     this.repaint,
     this.batchThreshold = 10, // 默认值为10，可以自行调整
   }) : _tick = tick,
